fix: satisfy fallow type refactor audit
